refactor(core): consolidate test-utils into a single module
TestHook, test_schema, InMemoryStorage, and test_provider were all
feature-gated on test-utils but scattered across runtime/hook.rs,
agent/tool.rs, repos/mem.rs, and model/test_provider.rs — each with
its own cfg(feature) stanza.

Move them into a single crates/core/src/test_utils/ module so the
feature has exactly one mount point. InMemoryStorage and
test_provider keep their file identities as submodules; TestHook and
test_schema move inline. External callers now import from
wcore::test_utils::* instead of four different paths.
